Move generation_utils into generation subfolder
