+++ optimum/graphcore/models/gpt2/modeling_gpt2.py
@@ -24,7 +24,7 @@ from transformers import GPT2ForSequenceClassification, GPT2ForTokenClassificati
 from transformers.modeling_outputs import CausalLMOutputWithCrossAttentions, SequenceClassifierOutputWithPast
 from transformers.models.gpt2.modeling_gpt2 import GPT2Attention
 
-from ...generation_utils import IPUGenerationMixin, _IndexedInputLinear
+from ...generation import IPUGenerationMixin
 from ...modeling_utils import (
     PipelineMixin,
     SerializedEmbedding,
